Avance inicial de la funcionalidad de login con JWT

+++ src/main/java/DSWS2Grupo4/repository/EmpleadoRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EmpleadoRepository extends JpaRepository<Empleado, Long> {
+public interface EmpleadoRepository extends JpaRepository<Empleado, Integer> {
     Optional<Empleado> findByUsername(String username);
 }
